reorganize update label func into bsl.go
